Move app domain events to their respective areas
