+++ Latte.Hotload/Program.cs
@@ -14,8 +14,6 @@ public static class Program
 
 	private static async Task Main( string[] args )
 	{
-		AppDomain.CurrentDomain.ProcessExit += OnProcessExit;
-		AppDomain.CurrentDomain.AssemblyResolve += ResolveAssembly;
 		CommandLineArguments = args.ToImmutableArray();
 
 		await AddAssemblyAsync( new AssemblyInfo
@@ -25,25 +23,9 @@ public static class Program
 		} );
 	}
 
-	private static Assembly? ResolveAssembly( object? sender, ResolveEventArgs args )
-	{
-		var assemblyName = args.Name[..args.Name.IndexOf( ',' )];
-		var assemblyPath = Path.GetFullPath( Path.Combine( "nuget", assemblyName + ".dll" ) );
-		if ( File.Exists( assemblyPath ) )
-			return Assembly.LoadFile( assemblyPath );
-
-		return null;
-	}
-
 	internal static async Task AddAssemblyAsync( AssemblyInfo assemblyInfo )
 	{
 		var assembly = HotloadableAssembly.New( assemblyInfo );
 		await assembly.InitAsync();
-	}
-
-	private static void OnProcessExit( object? sender, EventArgs e )
-	{
-		foreach ( var (_, assembly) in HotloadableAssembly.All )
-			assembly.Dispose();
 	}
 }
